+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.5.2] - 2024-05-19
 - Fix #40, add interrupt functions
+- update **MCP23S17_registers.h**
+- change return type of several functions
+  - e.g **bool enable/disableControlRegister()**
+- update readme.md
+- update keywords.txt
 
 
 ## [0.5.1] - 2024-03-02